BugFix: check whether `$APP_DIR/public/uploads` exists before creating symlink

+++ docker-entrypoint.sh
@@ -4,6 +4,8 @@ set -e
 
 # Corresponds to `FILE_UPLOAD=local`
 mkdir -p /data/uploads
-ln -s /data/uploads $APP_DIR/public/uploads
+if [ ! -e "$APP_DIR/public/uploads" ]; then
+  ln -s /data/uploads $APP_DIR/public/uploads
+fi
 
 exec "$@"
